requirements.txt
langchain
python-dotenv
pytesseract
Pillow
pypdfium2
streamlit
pandas
requests
openai